requirements.txt
mcp>=1.2.0
music21>=9.0.0
websocket>=15.0
